Rewordked the courseProjectTest file
Made changes to the drawing and setup of the file. As of this commit the draw method uses drawing methods from two different files just to test the functionality. If you want to test drawings make sure to change the path of the file to which the image is drawn.