Fixed text being drawn upside down
Fixed the issue with drawing text. Since we are transforming the coordinate system by flipping the y direction, the text was being drawn upside down so a flip had to occur along with drawing with a negative y in order to match with the coordinates in the new coordinate system.